Move reusable method to common api file
